LayoutTests/webaudio/audiochannelmerger-basic-expected.txt

Basic tests for AudioChannelMerger.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS IndexSizeError was thrown for numberOfInputs <= 0.
PASS IndexSizeError was thrown for numberOfInputs > 32.
PASS AudioChannelMerger created successfully with numberOfInputs = 32.
PASS AudioChannelMerger created successfully with empty parameter.
PASS ChannelMergerNode Object is available.
PASS successfullyParsed is true

TEST COMPLETE
